+++ task_sdk/pyproject.toml
@@ -28,6 +28,9 @@ dependencies = [
     "jinja2>=3.1.4",
     "methodtools>=0.4.7",
     "msgspec>=0.19.0",
+    'pendulum>=2.1.2,<4.0;python_version<"3.12"',
+    'pendulum>=3.0.0,<4.0;python_version>="3.12"',
+    "python-dateutil>=2.7.0",
     "psutil>=6.1.0",
     "structlog>=24.4.0",
     "retryhttp>=1.2.0,!=1.3.0",
